feat(infra): user/product/order 서비스에 Flyway 마이그레이션 내장
- spring-boot-starter-flyway와 flyway-database-postgresql 의존성 추가
- V1 스키마 SQL을 각 서비스 db/migration으로 이동, user는 private 스키마 명시
- 기본 비활성, prod config-repo에서만 spring.flyway.enabled 활성화
- Flyway 결과와 JPA 엔티티를 대조하는 FlywayMigrationValidationTest 추가
- 별도 마이그레이션 이미지(deploy/aws/database)와 미사용 V1_ 파일 삭제

+++ apps/product-service/src/main/resources/db/migration/V1__product_schema.sql
@@ -154,7 +154,8 @@ CREATE TABLE p_category_outbox_events (
 
 CREATE TABLE p_leaderboard_snapshots (
     id uuid PRIMARY KEY,
-    type integer NOT NULL,
+    -- LeaderboardType은 @Enumerated 없이 ORDINAL로 매핑되어 Hibernate가 smallint를 기대한다.
+    type smallint NOT NULL,
     target_id uuid NOT NULL,
     score double precision NOT NULL,
     ranking bigint NOT NULL,
